+++ src/components/index.ts
@@ -10,3 +10,4 @@ export * from "./Library";
 export * from "./Toast";
 export * from "./ErrorBoundary";
 export * from "./Skeletons";
+export * from "./KeyboardShortcuts";