+++ src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from "./tauriApi";
 export * from "./types";
 export * from "./constants";
+export * from "./hotkeys";